+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Output the filesystem type for the function `mount()`
+- Output the traces for the functions `clearenv()`, `getenv()`, `putenv()`,
+  `setenv()` and `unsetenv()`
 
 ### Fixed